<img width="364" alt="image" src="https://user-images.githubusercontent.com/356110/82735921-bbe63b00-9d25-11ea-97ef-ed4aee73ae43.png">

The core change is this https://github.com/multiformats/go-base36/blob/986b8a84d74dd7192b179abe4b77e1de66ab7549/base36.go#L50-L59 We now allocate just as much as we need, and iterate over the buffer in place: once to breakdown into values, and a second time to recode into an alphabet.


- before
```
go-multibase$ go test ./... -count=1 -bench .
goos: darwin
goarch: amd64
pkg: github.com/multiformats/go-multibase
BenchmarkRoundTrip/base36-8         	  326816	      3599 ns/op
BenchmarkRoundTrip/base36upper-8    	  332366	      3571 ns/op
BenchmarkRoundTrip/base58btc-8      	  481983	      2551 ns/op
BenchmarkRoundTrip/base58flickr-8   	  470001	      2556 ns/op
BenchmarkEncode/base36-8            	  435662	      2788 ns/op
BenchmarkEncode/base36upper-8       	  440614	      2788 ns/op
BenchmarkEncode/base58btc-8         	  770272	      1568 ns/op
BenchmarkEncode/base58flickr-8      	  700389	      1562 ns/op
BenchmarkDecode/base36-8            	 1606472	       745 ns/op
BenchmarkDecode/base36upper-8       	 1614858	       744 ns/op
BenchmarkDecode/base58btc-8         	 1245900	       952 ns/op
BenchmarkDecode/base58flickr-8      	 1259892	       956 ns/op
PASS
```

- after
```
go-multibase$ go test ./... -count=1 -bench .
goos: darwin
goarch: amd64
pkg: github.com/multiformats/go-multibase
BenchmarkRoundTrip/base36-8         	  432175	      2750 ns/op
BenchmarkRoundTrip/base36upper-8    	  436525	      2723 ns/op
BenchmarkRoundTrip/base58btc-8      	  466893	      2553 ns/op
BenchmarkRoundTrip/base58flickr-8   	  475454	      2547 ns/op
BenchmarkEncode/base36-8            	  601210	      1933 ns/op
BenchmarkEncode/base36upper-8       	  629655	      1938 ns/op
BenchmarkEncode/base58btc-8         	  791280	      1572 ns/op
BenchmarkEncode/base58flickr-8      	  795505	      1555 ns/op
BenchmarkDecode/base36-8            	 1558412	       766 ns/op
BenchmarkDecode/base36upper-8       	 1561417	       762 ns/op
BenchmarkDecode/base58btc-8         	 1250644	       954 ns/op
BenchmarkDecode/base58flickr-8      	 1263541	       954 ns/op
PASS
```

P.S. If you look closely you see that the decoding became marginally slower. To the best of my ability to reason it has to do with the *total* compiled object size. That is - if you **revert** https://github.com/multiformats/go-base36/commit/986b8a84d74, then the decoding number goes back to where it was, even though the reverted commit **does not touch the decode method** 🤯