@ribasushi 2020-06-02
👍 
I'll tag this as part of https://github.com/multiformats/go-multibase/issues/37 soon, want to let b58 bake a bit more...